requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# combine_images.py: 3
# util.py: 4
Pillow == 2.9.0.egg

# generate_map.py: 4,6
# read_log.py: 78,80
matplotlib == 1.4.2.egg

# combine_images.py: 5
# generate_map.py: 19
# model.py: 10
# train.py: 6
# util.py: 3
numpy == 1.11.3

# generate_map.py: 18
# train.py: 7
pandas == 0.18.1
tensorflow == 0.11.0

# generate_map.py: 22
# util.py: 1,2
scikit_image == 0.12.3

# combine_images.py: 8
scipy == 0.18.0

